Seperate my files into correct places
Error handler middleware is now the only thing that exists in the fil
Custom errors were moved to their own folder and each file will be its own error
Removed console.error from the catch blocks in the my user controller
