+++ backend-lift-log/src/app.ts
@@ -12,7 +12,7 @@ import programRouter from './routes/programRoutes';
 
 // Project Dependencies
 import { PORT } from './config';
-import { errorHandler } from './middleware/errors';
+import { errorHandler } from './middleware/errorHandler';
 
 
 const app = express();
@@ -20,25 +20,23 @@ const app = express();
 app.use(express.json());
 app.use(cors());
 
+
 // ROUTES
 app.use('/users', userRouter);
 app.use('/auth', authRouter);
 app.use('/programs', programRouter)
 
+// ERROR HANDLER MIDDLEWARE
+app.use(errorHandler)
+
 app.get('/', (req, res) => {
-    res.send('Hello World');
+    res.send('Lift Log API');
 });
 
 app.post("/test", async (req: Request, res: Response) => {
     res.send("Testing this for coverage")
 });
 
-/** Handle 404 errors -- this matches everything */
-// app.use(function (req, res, next) {
-//     throw new NotFoundError();
-// });
-
-app.use(errorHandler)
 
 app.listen(3000, () => {
     console.log(`Server is running on port ${PORT}`);
