+++ backend-lift-log/src/errors/UnauthorizedError.ts
@@ -1,0 +1,32 @@
+import { CustomError } from "../middleware/CustomError";
+
+export class UnauthorizedError extends CustomError {
+    private static readonly _statusCode = 400;
+    private readonly _code: number;
+    private readonly _logging: boolean;
+    private readonly _context: { [key: string]: any };
+
+    constructor(params?: { code?: number, message?: string, logging?: boolean, context?: { [key: string]: any } }) {
+        const { code, message, logging } = params || {};
+
+        super(message || "Unauthorized");
+        this._code = code || UnauthorizedError._statusCode;
+        this._logging = logging || false;
+        this._context = params?.context || {};
+
+        // Only because we are extending a built in class
+        Object.setPrototypeOf(this, UnauthorizedError.prototype);
+    }
+
+    get errors() {
+        return [{ message: this.message, context: this._context }];
+    }
+
+    get statusCode() {
+        return this._code;
+    }
+
+    get logging() {
+        return this._logging;
+    }
+}
