--- backend-lift-log/src/middleware/errors.ts
@@ -1,148 +1,0 @@
-"use strict";
-
-import { NextFunction, Request, Response } from "express";
-import { CustomError } from "./CustomError";
-
-export const errorHandler = (err: Error, req: Request, res: Response, next: NextFunction) => {
-    // Handled errors
-    if (err instanceof CustomError) {
-        const { statusCode, errors, logging } = err;
-        if (logging) {
-            console.error(JSON.stringify({
-                code: err.statusCode,
-                errors: err.errors,
-                stack: err.stack,
-            }, null, 2));
-        }
-
-        return res.status(statusCode).send({ errors });
-    }
-
-    // Unhandled errors
-    console.error(JSON.stringify(err, null, 2));
-    return res.status(500).send({ errors: [{ message: "Something went wrong" }] });
-};
-
-export class BadRequestError extends CustomError {
-    private static readonly _statusCode = 400;
-    private readonly _code: number;
-    private readonly _logging: boolean;
-    private readonly _context: { [key: string]: any };
-
-    constructor(params?: { code?: number, message?: string, logging?: boolean, context?: { [key: string]: any } }) {
-        const { code, message, logging } = params || {};
-
-        super(message || "Bad request");
-        this._code = code || BadRequestError._statusCode;
-        this._logging = logging || false;
-        this._context = params?.context || {};
-
-        // Only because we are extending a built in class
-        Object.setPrototypeOf(this, BadRequestError.prototype);
-    }
-
-    get errors() {
-        return [{ message: this.message, context: this._context }];
-    }
-
-    get statusCode() {
-        return this._code;
-    }
-
-    get logging() {
-        return this._logging;
-    }
-}
-
-export class UnauthorizedError extends CustomError {
-    private static readonly _statusCode = 400;
-    private readonly _code: number;
-    private readonly _logging: boolean;
-    private readonly _context: { [key: string]: any };
-
-    constructor(params?: { code?: number, message?: string, logging?: boolean, context?: { [key: string]: any } }) {
-        const { code, message, logging } = params || {};
-
-        super(message || "Unauthorized");
-        this._code = code || UnauthorizedError._statusCode;
-        this._logging = logging || false;
-        this._context = params?.context || {};
-
-        // Only because we are extending a built in class
-        Object.setPrototypeOf(this, UnauthorizedError.prototype);
-    }
-
-    get errors() {
-        return [{ message: this.message, context: this._context }];
-    }
-
-    get statusCode() {
-        return this._code;
-    }
-
-    get logging() {
-        return this._logging;
-    }
-}
-
-export class NotFoundError extends CustomError {
-    private static readonly _statusCode = 400;
-    private readonly _code: number;
-    private readonly _logging: boolean;
-    private readonly _context: { [key: string]: any };
-
-    constructor(params?: { code?: number, message?: string, logging?: boolean, context?: { [key: string]: any } }) {
-        const { code, message, logging } = params || {};
-
-        super(message || "Resource not found.");
-        this._code = code || NotFoundError._statusCode;
-        this._logging = logging || false;
-        this._context = params?.context || {};
-
-        // Only because we are extending a built in class
-        Object.setPrototypeOf(this, NotFoundError.prototype);
-    }
-
-    get errors() {
-        return [{ message: this.message, context: this._context }];
-    }
-
-    get statusCode() {
-        return this._code;
-    }
-
-    get logging() {
-        return this._logging;
-    }
-}
-
-export class ServerError extends CustomError {
-    private static readonly _statusCode = 400;
-    private readonly _code: number;
-    private readonly _logging: boolean;
-    private readonly _context: { [key: string]: any };
-
-    constructor(params?: { code?: number, message?: string, logging?: boolean, context?: { [key: string]: any } }) {
-        const { code, message, logging } = params || {};
-
-        super(message || "Server error, something went wrong.");
-        this._code = code || ServerError._statusCode;
-        this._logging = logging || false;
-        this._context = params?.context || {};
-
-        // Only because we are extending a built in class
-        Object.setPrototypeOf(this, ServerError.prototype);
-    }
-
-    get errors() {
-        return [{ message: this.message, context: this._context }];
-    }
-
-    get statusCode() {
-        return this._code;
-    }
-
-    get logging() {
-        return this._logging;
-    }
-}
